create logout widget
